Modified the way the content.js checks to see if the additional fields should be added.
Used to work by determining the user had an 'Unlock' symbol, indicating 
that they were elevated; but in the Dev Environment, we remove the need 
to elevate to an Admin in order to make Development simpler. 

New way checks to see if the update-set picker is there, as the 
additional elements involve interacting with update sets.